fix(api): path-blob width guard wrapped before it could fire on ILP32
pack_entries' cumulative-path-bytes guard (report L17) computed the sum
in size_t, which on 32-bit targets wraps before the 0xFFFFFFFF
comparison — -Werror=type-limits flagged the tautology on the armv7
leg, and the check was genuinely dead there. Widen to uint64.

+++ src/api/abi_contract.hpp
@@ -105,7 +105,10 @@ inline bool pack_entries(const std::vector<openrar::archive::BufferArchiveEntry>
     auto* out = reinterpret_cast<ArchiveEntryOut*>(entries.data());
     for (size_t i = 0; i < src.size(); ++i) {
         const auto& e = src[i];
-        if (paths.size() + e.path.size() > 0xFFFFFFFFull) return false;
+        // 64-bit intermediate: on ILP32 a size_t sum would wrap BEFORE the
+        // check, making the u32-width guard tautologically false (found by
+        // -Werror=type-limits on the armv7 leg).
+        if (static_cast<uint64_t>(paths.size()) + e.path.size() > 0xFFFFFFFFull) return false;
         out[i].path_offset = static_cast<uint32_t>(paths.size());
         out[i].path_len = static_cast<uint32_t>(e.path.size());
         out[i].is_dir = e.is_dir ? 1u : 0u;
